Introduction of instruction latency
- Hashmap from insn-name to its latency is introduced.
- At the time of reading the register, check if RAW dependency exists.
- If yes, and the value of source register is produced by a compute instruction, create a InsnLatencyEvent.
- If yes, and the value of source register is produced by a memory instruction, record the src reg and type in hashmap.
- At the time of writing the register, check if RAW dependency exists.
- If yes, set the dest reg id and type in the corresponding InsnLatency Event or in the hashmap.
- Change the elt interface for accessing vector registers, to include the access type.
- Load instruction upon acknowledging the reg should also set the latency of the register, if at all it was
  waiting for the load instruction result.

+++ riscv/mmu.cc
@@ -354,7 +354,7 @@ void mmu_t::register_memtracer(memtracer_t* t)
   tracer.hook(t);
 }
 
-void mmu_t::set_misses_dest_reg(uint8_t reg, spike_model::CacheRequest::RegType t)
+void mmu_t::set_misses_dest_reg(size_t reg, spike_model::CacheRequest::RegType t)
 {
   for(std::shared_ptr<spike_model::CacheRequest> miss: misses_last_inst)
   {
